flask-post-processing/requirements.txt
flask
Pillow
boto3
python-dotenv